api/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=1.8.2
email-validator>=1.1.3
python-dotenv>=0.19.0
starlette>=0.14.2
typing-extensions>=4.0.0
aiofiles>=0.7.0
bcrypt>=4.0.1
firebase-admin>=6.3.0